+++ logprep/run_logprep.py
@@ -28,7 +28,7 @@ from logprep.util.time_measurement import TimeMeasurement
 warnings.simplefilter("always", DeprecationWarning)
 logging.captureWarnings(True)
 
-DEFAULT_LOCATION_CONFIG = "/etc/logprep/pipeline.yml"
+DEFAULT_LOCATION_CONFIG = "file:///etc/logprep/pipeline.yml"
 getLogger("filelock").setLevel(ERROR)
 getLogger("urllib3.connectionpool").setLevel(ERROR)
 getLogger("elasticsearch").setLevel(ERROR)
@@ -170,7 +170,8 @@ def _verify_configuration(args, config, logger):
             config.verify_pipeline_only(logger)
         else:
             config.verify(logger)
-    except InvalidConfigurationError:
+    except InvalidConfigurationError as error:
+        logger.critical(error)
         sys.exit(1)
     except BaseException as error:  # pylint: disable=broad-except
         logger.exception(error)